Refactor symfony integration tests
